Core: Add SortedArray, which always keeps its elements sorted.
This was quite easy to do given the existing sort features in Array, and is
quite useful for cases where we need to run many queries for few changes.
Append is O(n), while Find is O(log(n)) (instead of resp. O(1) and O(n) with Array).

Also extend the Compare helpers to be able to query with a different type (for map-like behavior).