FBuildCore Protocol: add messages to get/set workers info and settings.
The blocking processes and grace period features are new and will be added in later commits.
These messages can be used to control the local worker (pause it while doing CI tests for example),
or remotely query the status of several workers (could be used for a distributed top-like dashboard).

On the Server side, the requested actions (get or set) are accumulated in the ServerState. New methods
GetRequestedServerInfoLevel() and GetRequestedChanges() needs to be called to retrieve and process them.