Second attempt at changing build tasks priorities to minimize waiting for links to complete at the end.
Recursive cost computation was only taking into account the "earliest" dependency chain for each node, because node traversal is done only once for each node. For projects where all targets are a dependency of the "ALL" default target (such as CMake projects), this meant that everything had the same cost. Instead, when we visit a node that was already traversed but for which we have an increased cost, then we recursively update its and its parents' cost (which is much cheaper that a "full" traversal updating statuses and other info).

DLL and Library nodes have the highest priority (because then can lead to long chains of sequential links), distributed nodes the lowest (because other workers can handle them while the local node focuses on local-only tasks).